+++ packages/article-skeleton/src/styles/article-body/responsive.web.js
@@ -1,5 +1,5 @@
 import { View } from "react-native";
-import styled from "styled-components";
+import styled, { css } from "styled-components";
 import {
   breakpoints,
   colours,
@@ -42,21 +42,27 @@ export const Paragraph = styled.p`
 `;
 
 export const InteractiveContainer = styled.div`
-  padding-left: ${spacing(2)};
-  padding-right: ${spacing(2)};
   margin-bottom: ${spacing(4)};
 
-  @media (min-width: ${breakpoints.medium}px) {
-    width: 80.8%;
-    margin: 0 auto;
-    padding-left: 0;
-    padding-right: 0;
-    margin-bottom: ${spacing(4)};
-  }
+  ${props =>
+    props.fullWidth
+      ? css`
+          width: 100%;
+        `
+      : css`
+          padding: 0 ${spacing(2)};
 
-  @media (min-width: ${breakpoints.wide}px) {
-    width: 56.2%;
-  }
+          @media (min-width: ${breakpoints.medium}px) {
+            width: 80.8%;
+            margin-left: auto;
+            margin-right: auto;
+            padding: 0;
+          }
+
+          @media (min-width: ${breakpoints.wide}px) {
+            width: 56.2%;
+          }
+        `};
 `;
 
 /* --- Lead Asset Styles --- */