+++ packages/edition-slices/src/slices/leadoneandone/index.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import PropTypes from "prop-types";
 import { LeadOneAndOneSlice } from "@times-components/slice-layout";
-import { TileA, TileB, TileU, TileAA } from "../../tiles";
+import { TileA, TileB, TileC, TileU, TileAA, TileAQ } from "../../tiles";
 import { ResponsiveSlice } from "../shared";
 
 class LeadOneAndOne extends Component {
@@ -9,6 +9,8 @@ class LeadOneAndOne extends Component {
     super(props);
     this.renderSmall = this.renderSmall.bind(this);
     this.renderMedium = this.renderMedium.bind(this);
+    this.renderWide = this.renderWide.bind(this);
+    this.renderHuge = this.renderHuge.bind(this);
   }
 
   renderSmall(breakpoint) {
@@ -41,14 +43,49 @@ class LeadOneAndOne extends Component {
             tileName="lead"
           />
         }
-        support={
-          <TileAA
+        support={<TileAA onPress={onPress} tile={support} tileName="support" />}
+      />
+    );
+  }
+
+  renderWide(breakpoint) {
+    const {
+      onPress,
+      slice: { lead, support }
+    } = this.props;
+    return (
+      <LeadOneAndOneSlice
+        breakpoint={breakpoint}
+        lead={
+          <TileU
             breakpoint={breakpoint}
             onPress={onPress}
-            tile={support}
-            tileName="support"
+            tile={lead}
+            tileName="lead"
           />
         }
+        support={<TileC onPress={onPress} tile={support} tileName="support" />}
+      />
+    );
+  }
+
+  renderHuge(breakpoint) {
+    const {
+      onPress,
+      slice: { lead, support }
+    } = this.props;
+    return (
+      <LeadOneAndOneSlice
+        breakpoint={breakpoint}
+        lead={
+          <TileU
+            breakpoint={breakpoint}
+            onPress={onPress}
+            tile={lead}
+            tileName="lead"
+          />
+        }
+        support={<TileAQ onPress={onPress} tile={support} tileName="support" />}
       />
     );
   }
@@ -58,8 +95,8 @@ class LeadOneAndOne extends Component {
       <ResponsiveSlice
         renderSmall={this.renderSmall}
         renderMedium={this.renderMedium}
-        renderWide={this.renderMedium}
-        renderHuge={this.renderMedium}
+        renderWide={this.renderWide}
+        renderHuge={this.renderHuge}
       />
     );
   }