+++ packages/edition-slices/src/tiles/tile-u/styles/index.js
@@ -1,10 +1,30 @@
 import {
-  fonts,
+  fontFactory,
   spacing,
   editionBreakpoints
 } from "@times-components/styleguide";
 
+const sharedStyles = {
+  container: {
+    flex: 1,
+    flexDirection: "row"
+  },
+  headline: {
+    ...fontFactory({
+      font: "headline",
+      fontSize: "tileLeadHeadline"
+    })
+  },
+  imageContainer: {
+    width: "60%"
+  },
+  summaryContainer: {
+    width: "40%"
+  }
+};
+
 const mediumBreakpointStyles = {
+  ...sharedStyles,
   container: {
     flex: 1,
     paddingVertical: spacing(3),
@@ -18,7 +38,7 @@ const mediumBreakpointStyles = {
     paddingBottom: spacing(2)
   },
   headline: {
-    fontFamily: fonts.headline,
+    ...sharedStyles.headline,
     fontSize: 40,
     lineHeight: 40,
     marginBottom: spacing(2)
@@ -26,14 +46,36 @@ const mediumBreakpointStyles = {
 };
 
 const wideBreakpointStyles = {
-  ...mediumBreakpointStyles,
+  ...sharedStyles,
+  container: {
+    ...sharedStyles.container,
+    padding: spacing(2)
+  },
   headline: {
-    ...mediumBreakpointStyles.headline,
-    marginBottom: 0
+    ...sharedStyles.headline,
+    paddingVertical: spacing(1)
+  },
+  summaryContainer: {
+    ...sharedStyles.summaryContainer,
+    paddingRight: spacing(4)
   }
 };
 
-export default breakpoint =>
-  breakpoint === editionBreakpoints.medium
-    ? mediumBreakpointStyles
-    : wideBreakpointStyles;
+const hugeBreakpointStyles = {
+  ...wideBreakpointStyles,
+  headline: {
+    ...wideBreakpointStyles.headline,
+    ...fontFactory({
+      font: "headline",
+      fontSize: "articleHeadline"
+    })
+  }
+};
+
+const stylesResolver = {
+  [editionBreakpoints.medium]: mediumBreakpointStyles,
+  [editionBreakpoints.wide]: wideBreakpointStyles,
+  [editionBreakpoints.huge]: hugeBreakpointStyles
+};
+
+export default breakpoint => stylesResolver[breakpoint];