+++ packages/provider-queries/src/author-articles-no-images.js
@@ -1,38 +1,7 @@
 import { addTypenameToDocument } from "apollo-utilities";
-import gql from "graphql-tag";
+import authorArticlesNoImagesQuery from "./author_articles_no_images.graphql";
 
-export default addTypenameToDocument(gql`
-  query ArticleListQuery(
-    $first: Int
-    $longSummaryLength: Int
-    $shortSummaryLength: Int
-    $skip: Int
-    $slug: Slug!
-  ) {
-    author(slug: $slug) {
-      articles {
-        count
-        list(first: $first, skip: $skip) {
-          hasVideo
-          id
-          label
-          slug
-          shortIdentifier
-          leadAsset {
-            __typename
-          }
-          longSummary: summary(maxCharCount: $longSummaryLength)
-          publicationName
-          publishedTime
-          headline
-          shortHeadline
-          shortSummary: summary(maxCharCount: $shortSummaryLength)
-          url
-        }
-      }
-    }
-  }
-`);
+export default addTypenameToDocument(authorArticlesNoImagesQuery);
 
 export const propsToVariables = ({
   longSummaryLength,