+++ packages/button/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.103](https://github.com/newsuk/times-components/compare/@times-components/button@2.3.102...@times-components/button@2.3.103) (2019-08-26)
+
+**Note:** Version bump only for package @times-components/button
+
+
+
+
+
 ## [2.3.102](https://github.com/newsuk/times-components/compare/@times-components/button@2.3.101...@times-components/button@2.3.102) (2019-08-09)
 
 **Note:** Version bump only for package @times-components/button