+++ packages/provider-test-tools/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.17.37](https://github.com/newsuk/times-components/compare/@times-components/provider-test-tools@1.17.36...@times-components/provider-test-tools@1.17.37) (2019-08-26)
+
+**Note:** Version bump only for package @times-components/provider-test-tools
+
+
+
+
+
 ## [1.17.36](https://github.com/newsuk/times-components/compare/@times-components/provider-test-tools@1.17.35...@times-components/provider-test-tools@1.17.36) (2019-08-14)
 
 **Note:** Version bump only for package @times-components/provider-test-tools